+++ main.go
@@ -112,4 +112,31 @@ func main() {
 	}
 	testLogger(logger3)
 
+	// ------------------------------------------------------------------------
+	// README.md examples
+	// ------------------------------------------------------------------------
+
+	var (
+		ComponentId = 9999            // See https://github.com/Senzing/knowledge-base/blob/main/lists/senzing-component-ids.md
+		IdMessages  = map[int]string{ // Message templates. Example: https://github.com/Senzing/init-database/blob/main/senzingconfig/main.go
+			2000: "Today's greeting:  %s",
+			4000: "Here's what happened: %s",
+		}
+		callerSkip = 3 // Used to determine "location" information. See https://pkg.go.dev/runtime#Caller
+	)
+
+	// Logging options. See https://github.com/Senzing/go-logging/blob/main/logging/main.go
+	loggerOptions := []interface{}{
+		&logging.OptionCallerSkip{Value: callerSkip},
+	}
+
+	logger, err := logging.NewSenzingToolsLogger(ComponentId, IdMessages, loggerOptions...)
+	if err != nil {
+		fmt.Println(err)
+	}
+
+	logger.Log(2000, "Hello, world!")
+
+	err = logger.Error(4000, "A bad thing")
+	fmt.Printf("The error: %v\n", err)
 }