+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [markdownlint](https://dlaa.me/markdownlint/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.4] - 2023-05-23
+
+### Changed in 1.2.4
+
+- Added documentation on `Log()`, `Error()`, and logging output.
+- Removed a layer of "caller depth" in `logging.Error()`
+
 ## [1.2.3] - 2023-05-11
 
 ### Changed in 1.2.3